* ability to cache images locally * ability to specify application executable name

+++ src/notifier.test.ts
@@ -8,7 +8,7 @@ import {
 } from '@jest/globals';
 import { createNotifier } from './createNotifier.js';
 import { StatusMessageType } from './statusMessageType.js';
-import { Notifier } from './notifier.js';
+import { Notifier, NOTIFIER_OPTIONS } from "./notifier.js";
 import { Notification } from './notification.js';
 import { exec } from 'child_process';
 import { StatusMessage } from './statusMessage.js';
@@ -22,20 +22,24 @@ describe('notifications manual tests', () => {
     let notification: Notification;
     let sysNotification: Notification;
     let notificationTemplate: string;
+    let apiKey: string | undefined = undefined;
     beforeEach(async () => {
         notifier = await createNotifier({
             application_id: 'notifier-test', // use process.execPath after start menu fix
-            connectToExistingService: false,
+            connectToExistingService: Boolean(apiKey),
+            api_key: apiKey,
             port: 7070,
         });
     });
     afterEach(async () => {
         await sysNotification?.remove();
-        await notification.remove();
+        await notification?.remove();
         await notifier.close();
     });
     afterAll(() => {
-        exec(`taskkill /f /im win-toast-notifier.exe`);
+        if (!apiKey){
+            exec(`taskkill /f /im win-toast-notifier.exe`);
+        }
     });
 
     async function createNotification(msg: string) {
@@ -205,5 +209,126 @@ describe('notifications manual tests', () => {
             expect(status.info?.inputs['vars']).toBe('yep');
             await notification.remove();
         });
+    });
+    describe('sandbox', () => {
+        test('simple', async () => {
+            notification = await notifier.notify({
+                body: 'Hello',
+                audio: {src: NotificationSounds.SMS},
+                actions: [
+                    {
+                        actionType: 'input',
+                        type: 'selection',
+                        id: 'selectOptions',
+                        defaultInput: 'option1',
+                        selection: [
+                            {
+                                id: 'option1',
+                                content: 'Option 1'
+                            },
+                            {
+                                id: 'option2',
+                                content: 'Option 2'
+                            }
+                        ]
+                    },
+                    {
+                        actionType: 'action',
+                        content: 'Button 1',
+                        arguments: 'button1Pressed'
+                    },
+                    {
+                        actionType: 'action',
+                        content: 'Button 2',
+                        arguments: 'button1Pressed'
+                    }
+                ]
+            });
+            await new Promise<void>((cb) => {
+                notification.onChange(statusMessage => {
+                    if (statusMessage.type == StatusMessageType.Activated){
+                        if (statusMessage.info?.arguments === 'button1Pressed'){
+                            console.log(`You pressed button 1`);
+                            cb();
+                        }
+                    }
+                });
+            });
+        });
+        test('complex notification', async () => {
+            NOTIFIER_OPTIONS.base_path += "/..";
+            let imageUrl = 'https://gravatar.com/avatar/bd57abc1c5b0dd6a478fd9bb05fa62be?s=400&d=robohash&r=x';
+            notification = await notifier.notify({
+                settings: {image_cache: {enable: true}},
+                audio: {
+                    src: NotificationSounds.SMS
+                },
+                body: [
+                    {
+                        type: 'image',
+                        'hint-crop': 'circle',
+                        src: imageUrl,
+                        placement: 'appLogoOverride'
+                    },
+                    {
+                        content: 'admin mentioned you in a comment',
+                        type: 'text'
+                    },
+                    {
+                        content: '@Test User hello!',
+                        type: 'text'
+                    }
+                ],
+                actions: [
+                    {
+                        id: 'quickReply',
+                        actionType: 'input',
+                        type: 'selection',
+                        defaultInput: 'option2',
+                        selection: [
+                            {
+                                id: 'option1',
+                                content: 'Option 1'
+                            },
+                            {
+                                id: 'option2',
+                                content: 'Option 2'
+                            },
+                            {
+                                id: 'option3',
+                                content: 'Option 3'
+                            }
+                        ]
+                    },
+                    {
+                        id: 'replyText',
+                        actionType: 'input',
+                        type: 'text'
+                    },
+                    {
+                        actionType: 'action',
+                        content: 'Reply',
+                        arguments: 'sendReply',
+                        "hint-inputId": 'replyText'
+                    },
+                    {
+                        actionType: 'action',
+                        content: 'Open',
+                        arguments: 'open',
+                    },
+                    {
+                        actionType: 'action',
+                        content: 'Dismiss',
+                        arguments: 'dismiss',
+                    }
+                ]
+            });
+            await new Promise<void>((res) => {
+                notification.onChange(statusMessage => {
+                    console.log(statusMessage.type);
+                    res();
+                })
+            })
+        })
     });
 });
