+++ src/notifier.ts
@@ -4,26 +4,53 @@ import { NotifierSettings } from './notifierSettings.js';
 import path from 'path';
 import fetch, { HeadersInit } from 'node-fetch';
 import { StatusMessage } from './statusMessage.js';
-import { NotificationConfig } from './notification-config.js';
+import {
+    NotificationConfig,
+    NotificationImage,
+} from './notification-config.js';
 import { buildNotificationXml } from './toastXmlBuilder.js';
-import {fileURLToPath} from "url";
+import { fileURLToPath } from 'url';
+import { downloadFile } from './network-utils.js';
 
 interface NotifierConfig {
     application_id: string;
     api_key: string;
     ip: string;
     port?: number;
+    connectedToExistingService: boolean;
 }
 
 interface NotifyResponse {
     id: string;
 }
 
+export const NOTIFIER_OPTIONS = {
+    base_path: path.dirname(fileURLToPath(import.meta.url)),
+    image_cache: {
+        _folder_path: '',
+        folder_name: 'image-cache',
+        set folder_path(value: string){
+            this._folder_path = value;
+        },
+        get folder_path() {
+            return this._folder_path || path.resolve(
+              NOTIFIER_OPTIONS.base_path,
+              this.folder_name
+            );
+        }
+    },
+};
+
 export class Notifier {
-    public static BinaryPath: string = path.resolve(
-        path.dirname(fileURLToPath(import.meta.url)),
-        '../bin/win-toast-notifier.exe'
-    );
+    public static ExecutableName: string = 'win-toast-notifier.exe';
+
+    public static get BinaryPath(): string {
+        return path.resolve(
+            path.dirname(fileURLToPath(import.meta.url)),
+            `../bin/${Notifier.ExecutableName}`
+        );
+    }
+
     private _process?: ChildProcessWithoutNullStreams;
     private _onReady?: Function;
     private readonly _onReadyPromise: Promise<void>;
@@ -31,6 +58,7 @@ export class Notifier {
         api_key: '1',
         ip: '127.0.0.1',
         application_id: '',
+        connectedToExistingService: false,
     };
     private _notifications: Map<string, Notification> = new Map<
         string,
@@ -47,6 +75,7 @@ export class Notifier {
             settings.application_id ?? this._config.application_id;
         this._config.ip = settings.ip ?? this._config.ip;
         if (settings.connectToExistingService) {
+            this._config.connectedToExistingService = true;
             this._subscribeForEvents();
             this._onReady?.();
             return;
@@ -143,17 +172,42 @@ export class Notifier {
     }
 
     async close() {
-        const result = await fetch(this._getUrl('quit'), {
-            headers: this._getHeaders(),
-        });
-        const response = await result.text();
-        console.debug(`App quit: ${response}`);
-        this._process?.kill('SIGTERM');
+        if (!this._config.connectedToExistingService) {
+            const result = await fetch(this._getUrl('quit'), {
+                headers: this._getHeaders(),
+            });
+            const response = await result.text();
+            console.debug(`App quit: ${response}`);
+            this._process?.kill('SIGTERM');
+        }
     }
 
     async notify(config: NotificationConfig) {
+        if (config.settings?.image_cache?.enable) {
+            config = await this._precacheImages(config);
+        }
         let xml = buildNotificationXml(config);
         return await this.notifyRaw(xml);
+    }
+
+    async _precacheImages(config: NotificationConfig) {
+        const body = config.body;
+        if (!body.length) {
+            return config;
+        }
+        const result = { ...config };
+        for (const item of body) {
+            const image = item as NotificationImage;
+            if (image.type === 'image' && image.src?.startsWith('http')) {
+                image.src = await downloadFile(
+                    image.src,
+                    NOTIFIER_OPTIONS.image_cache.folder_path,
+                    config.settings?.image_cache?.expiration_timeout_seconds ??
+                        5 * 60
+                );
+            }
+        }
+        return result;
     }
 
     async notifyRaw(xml: string) {
@@ -166,6 +220,14 @@ export class Notifier {
             headers: this._getHeaders(),
         };
         let response = await fetch(url, config);
+        if (response.status === 401) {
+            if (this._config.connectedToExistingService) {
+                throw `You specified connectToExistingService: true, provide valid api_key, current api_key: ${this._config.api_key}`;
+            }
+        }
+        if (!response.ok) {
+            throw response.statusText;
+        }
         let response_json = (await response.json()) as NotifyResponse;
         let notification = new Notification(this, response_json.id);
         this._notifications.set(notification.id, notification);
